Updated info for Comfort [Inn|Suites]
Addresses #1843 and #1841

This is a little tricky because Comfort Inn & Suites does not officially exist
anymore. According to Choice Hotels (the parent company) Comfort Inn and Comfort
Suites are two distinct brands. For this reason I added each one to the other's
"nomatch" tag. This might not be the right call.

I also removed the entry for "Comfort Inn & Suites" but set it to match with
"Comfort Inn" because if you go to the Choice Hotels page for "Comfort Inn &
Suites" you are redirected to the "Comfort Inn" reservation page.