update (#2)
* Correct RadioShack to be one word

The brand is one word. Even the wikipedia article name included already is one word.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Cabela's

Popular hunting / fishing store in the US / Canada.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Bass Pro Shops

Popular hunting / fishing store in the US / Canada.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add US Cellular mobile stores

They're a big carrier in the areas of the US that ATT / Verizon don't cover well.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Veggie Grill

A vegan fast food restaurant on the west coat of the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Sur La Table

High end kitchen store based out of Seatle with locations throughout the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add wikipedia links for Watsons and Walgreens

Might as well add these.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Express wikidata and suppress warnings

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix count warning with Dutch Bros

Once this preset is in I'm going to get a task setup to fix every single one of the existing incorrectly tagged names.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix more npm build warnings

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add BevMo! liquor store

This is a west coast. US liquor store chain.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Ashley HomeStore

This is all over OSM, but it's tagged as Ashley Furniture HomeStore.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Restoration Hardware

Popular high end furniture store in the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Mud Bay pet store

A pacifiic northwest US pet store that's been expanding rapidly in the last few years.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Paper Source

Fancy stationary store in the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add BJ's Wholesale

A Wholesale club on the east coast of the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Pick 'n Save

Wisconsin supermarket chain run by Roundy's.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Aaron's wikipedia / wikidata tags

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add a few more wikipedia entries

Filling in the blanks

Signed-off-by: Tim Smith <tsmith@chef.io>

* Resolve warnings

* npm run build

* Nando's cuisine can be "chicken;portuguese"
(closes #2272)

* Społem can be both a supermarket or a convenience store
(closes #2269)

* Add wikidata for Spar supermarket/convenience stores

* Add the Vitamin Shoppe

Seems to exist in just about every mall in the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Tuesday Morning

This store is all over the map in OSM so I added a bunch of matches here. It's like Ross for your house so housewares seems like the best match.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add TCBY

Popular yogurt place that seems to exist in every mall.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Ann Taylor

Popular high end clothing store part of Ann Inc.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Mattress Warehouse

This one is odd. There's a large number of these stores on the east coast in the US, but for some reason there's no wikipedia article. Apparently mattresses aren't cool to write about.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add PacSun

Another popular clothing brand that seems to have a store in every mall in the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Origins

A popular Cosmetics company found in many malls in the US.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add Wikipedia entries