Update the counts
In doing so new brands were added and several old ones left as they dropped below 50. For the old ones I either added a match or added the brand with data.

Spar Express dropped off as a supermarket since it's really a convenience store. I'm adding this as a match for the entry we already have.
The electronics shop Mediamarkt dropped off, but I created another match for the proper case + spacing of the name, which is what this entry should have been the whole time
The bank Banco Union was added as another match for Banco Unión
The bank Caisse d'épargne was added as a match for Caisse d'Épargne
Chata Polska dropped off the list, but I created a wikidata page for it and added those details
Johnny Rockets dropped off so I added it. These are US style
Pret a Manger was added as a match to Pret A Manger with a capital A which is the correct variant.
Кулинария convenience stores and cafes have been removed as it was a generic and it dropped off the list for cafes. It means "cooking" in Russian.
GPL fuel dropped off, but I think that was just a station selling LPG and not a brand
La Mie Caline dropped below 50 and became a match on La Mie Câline
Ulta dropped below 50 and became a match on Ulta Beauty

Signed-off-by: Tim Smith <tsmith@chef.io>