Vending machines in Germany: Add details
Packstation (self-service collection and dispatch of parcels):
"Packstation" is the brand name of this kind of machine, while "DHL" is
the brand of the operator.

"Paketbox" is a similar machine by the same operator under a different
brand where you can only send parcels.

Add vending tags to some more entries.

Fixes #959