add necessary nomatch to Tesco brands and added wikidata (#2521)
Stand-alone Tesco petrol stations usually (always in my experience) have the petrol station branded as Esso (a separate fuel company) and the shop branded as Tesco Express, but the petrol stations that are attached to larger stores have both branded as just Tesco, including the separate convenience store that is not attached to the larger store. A brief look at the existing tags on Overpass indicated that shop=convenience name=Tesco was mostly being used for this and not incorrect tagging of Tesco Expresses, though I may just have been lucky with the shops I looked at.