Fix Co-op bug
The Co-op is actually the brand in Canada so I add countryCodes and add nomatch tag to not conflict with the UK The Co-operative. While, in "The Co-operative" tag which is real Co-op in Uk, I add more match tag to receive the "Co-op" search and, similiarly, add more nomatch tag.