Split out Bonobo and Bonobos (#2609)
Split out Bonobo and Bonobos

Bonobo is a retailer in France, but we gave it wikidata for Bonobos which is a retailer in the US. Split out these two entries and give them country codes. I created a wikidata page for Bonobo, but I'm not having luck finding the necessary information to fill it in. I'd appreciate help from someone in France who might have better luck searching for their Twitter / Facebook account / anything.

Signed-off-by: Tim Smith <tsmith@chef.io>