Add country codes to several variety stores
Dollar Tree is in the US and Canada
Family Dollar is only in the US
Home Bargains is in the UK. There's a single mapped location in India,
but the website has a list of all stores and that's not one of them. Not
sure about that.
Poundworld is in UK/Ireland

Signed-off-by: Tim Smith <tsmith@chef.io>