Creating police section for major Canadian police forces
The RCMP (Royal Canadian Mounted Police) is the police force in most of Canada, except for Ontario and Quebec, which both have their own provincial police forces. Edits were made in the online client, so apologies in advance for possible syntax errors. Getting back on this project after a while away.