Added Biscuitville as mentioned in issue #3368
I have added the Biscucitville brand to brands/amenity/fast_food.json. As per the information available on the and from Biscuitville's offical website, the company is primarily involved in selling biscuits made from their own recipe. It was involved in the pizza and bread business but biscuits is what they are known for. So, "cuisine" is given the value of "biscuit". Other better options can be "cuisine" : "snacks".