Fix Sports Direct pointing to parent organisation
Bit of a weird one since the parent organisation is Sports Direct
International which started with Sports Direct. However, the
organisation now owns many different brands and this other item is
specifically for the Sports Direct brand. Not the end of the world since
we know stores tagged with the old brand:wikidata are actually this
store.