Generate identifiers that conform to /^\w+$/ to avoid unicode or right-to-left surprises. Also split out the id generation into its own function. (re: #3995)
This now exposed a bunch of brands that were duplicate in the index
so the duplicates have been removed.
(e.g. an item in English and an item in Thai for the same thing, but
both with same `name:en` and same locationSet)

The duplicate removal just happened to close #4106 also.