Update mall.json
Scentre group is the operator, but the Westfield name and brand is what appears on the buildings.  Not sure how to tag the brand as there isn't a specific Westfield Australia entry anywhere.

Remove name from GPT and Vicinity as the malls have individual names.

Fix copy paste error with Vicinity brand