remove dubious entries for Polish fire station operators
KM PSP <place name> is merely the name of the local command of Państwowa Straż Pożarna, the state-run professional firefighting organisation, which is the most popular operator name in this category on Poland.

ZOSP RP is not an operator of fire stations but a union of non-public associations, which themselves operate volunteer fire brigades.