Added secure SSL check to nsi.guide
Added a small check to each NSI entry to see if the URL listed on the Wikidata is secure or not. If it isn't, it's highlighted in red.

Not all web sites are SSL these days, but this little addition invites people to check the Wikidata to see if an SSL version of the web site is available, and to update the Wikidata if it is. #4717