Use Special:FilePath for .svg instead of generating our own urls (closes #5928)
The P145 claim on Q6109362 contained 'wetherspoon logo.svg' (lowercase)
but the actual url should be 'Wetherspoon logo.svg' (uppercase)

Apparently this mismatch can happen sometimes because the filenames get
normalized, but claims aren't updated to match the true filename.

We can use Special:FilePath to redirect the user to the real file.
Thanks @jleedev for the solution!

https://osmus.slack.com/archives/CUP8V1Z61/p1641313622080700