"Deutsche Telekom" doesn't operate any emergency=phone.
The German company "Deutsche Telekom" does not operate any emergency=phone. Those emergency=phone's which are tagged with this operator (according to taginfo there exist 62 ones) are misinterpreted as emergency-phones, but they are normal public telephones (amenity=telephone). So there shouldn't be an operator-ing in the nsi for this. See also my comment in https://github.com/osmlab/name-suggestion-index/pull/6177#issuecomment-1042733139