Excluded South Vietnamese flag from Vietnam
The historical flag of the Republic of Vietnam is not only rare but probably illegal to fly in Vietnam. South Vietnam and Vietnam share the same country code, so by coincidence, iD has been preferring the South Vietnamese flag over the Vietnamese flag, which is probably realistic outside of Vietnam but not within the country.

Also tagged the Vietnamese flag with its color.

Fixes #6223.