Change "Deutsche Telekom" charging station operator to the consistent value "Deutsche Telekom AG"
For the operator "Deutsche Telekom [AG]" in Germany, we should use a consistent value.
We do have "operator=Deutsche Telekom AG" (+ operator:wikidata=Q9396) already in the NSI for amenity=telephone and telecom=exchange. So we should use the value "Deutsche Telekom AG" for charging stations, too, to make it consistent. Especially because the operator for the telephones, the telephone exchanges and the charging stations is one adn the same, here.