Western Power Distribution -> National Grid Electricity Distribution
Western Power Distribution (which is a power distribution network
operator in England) has been acquired by National Grid and recently
rebranded to "National Grid Electricity Distribution". This changeset
removes the old operator and adds the new one.

The OSM-GB community has decided that it's best to keep this separate
from "National Grid" as a whole (which also operates a different part
of the electricity network in the UK).

This changeset also removes "Western Power" (which I believe was an
error/duplicate) and "Western Power Distribution (South West) plc"
(which was/is an internal division of WPD/NGED - we've decided to put
this distinction in the `owner` tag).