Clean up UK power operators
This changeset cleans up the list of UK power network operators, adding
wikidata and more specific country codes where possible. This should now
(hopefully) line up with the list of operators here:

https://www.energynetworks.org/customers/find-my-network-operator

"BritNed" has been removed as it only operates one interconnector and is
probably not worth having in the NSI.

"EEB", "NEDL", and "YEDL" have been removed as they are historic region
names which are now part of other operators.

Manx Utilities is also added for the Isle of Man.