Update charging_station.json
Updates to reflect Z energy has its own brand and is no longer using Openloop.
Additionally have removed duplicate Openloop entries in operators.
Openloop is now a 'Brand', but have left entries for counties energy and wel networks as operator with network as openloop. - So stations can now be tagged as either openloop as brand, or counties energy/ wel networks as operator and openloop as network.