Avoid toLowerCasing İ (\u0130) - it changes (closes #8261)
This was tricky because you'd think that case insensitive regex /i would catch
both upper and lower case variants of this, but it doesn't.

Then, I tried to match both variants with a regex like '^(İ|i̇)nşaat malları$',
but toLowerCasing *that* regex in our file_tree writing code was changing the 'İ'.

So for now, our build scripts can just avoid toLowerCasing a string with a 'İ' in it.