Add Vision Source
Vision Source[^1] is a network of independently-owned and operated eye
doctors in the United States. For the purposes of OSM, it functions as a
brand. The Vision Source logo is (usually) prominently featured on the
signage for member practices, if you buy glasses they come in a Vision
Source-branded plastic bag, and most member practices use the same
website theme with the same marketing blog posts.

[^1]: https://www.visionsource.com; https://en.wikipedia.org/wiki/Vision_Source