remove Flock Safety from operators
Flock is not an operator as defined in the OSM wiki ("directly in charge of the current operation of a map object"), since Flock's customers are the operators of their cameras. Their customers are the ones who instruct the device to alert on specific license plates, conduct searches of the data, etc. Flock just leases these devices to their customers and provides a cloud platform to access them.