Change the Wikidata item for A+ Glass (#10773)
* Change the Wikidata item for A+ Glass

Wikidata item Q123409768 is for a shop named "Active Green + Ross",
located in Canada. It is already linked to the correct brand in NSI,
with ID activegreenross-9c9847.

Item Q116688243 matches this brand's name and region.

Co-authored-by: Snowysauce <22599992+Snowysauce@users.noreply.github.com>

* Rename A+ Glass to A+Glass

The company writes its brand name without a space.

While most entries in OSM spell it with a space, the majority of edits
are from a single user.

In the absence of a clear consensus, we chose to follow the brand's own
spelling.

---------

Co-authored-by: Snowysauce <22599992+Snowysauce@users.noreply.github.com>