add New Zealand Territorial Authority districts
These locationSets represent the approximate boundaries of NZ
territorial authorities. These are distinct from the approximate regions
described by ISO 3166-2:NZ. My ultimate goal is to be able to, for
example, recommend that tags be upgraded for substations within the
Christchurch City territorial authority boundary be upgraded to reflect
that the owner/operator is Orion New Zealand Limited.