Let's give bun a try
Bun is a new JavaScript/TypeScript runtime with everything built in.
It seems like it could be a lot easier than using a bunch of tools.
It also is a bit faster and has some improvements to security.
This replaces node, c8, esbuild, taze, maybe some other things.

This commit includes the initial converstion to use Bun.
These things work:
- building the geojsons and index
- running lint, tests

Todo:
- nsi.guide and the local server
- the wikidata sync script
- building dist