Simplify cleanup steps
Removing everything under `./dist` would be easiest, but that throws
away the wikidata outputs, which take a long time to generate.

So we will put them in their own folder that only gets refreshed
when the user runs the `bun run wikidata` script.

These are the places where dist files will be placed:

./dist/js        (javascript bundles, nsiguide - `bun run build:js`)
./dist/ts        (typescript definition files - `bun run build:ts`)
./dist/json      (most json files - `bun run build:json`)
./dist/presets   (iD and JOSM presets  - `bun run dist`)
./dist/wikidata  (wikidata sync files - `bun run wikidata`)