chore: remove some brand not really work in China
At least they don't have much branch office and enough to set as a separated entry in China

I think we shouldn't confused commercial bank and investment bank in this entry. investment bank should be seen as office not bank I guess.

For UOB, it really have business in China so I only separeated it with localized entry now, but may make further cleanup in future.